# improve performance of partition skeleton a bit