On `main`:
```julia
julia> using RootedTrees, BSeries, BenchmarkTools, StaticArrays

julia> let order = 10
           # explicit midpoint method
           A = @SArray [0 0; 1//2 0]; b = @SArray [0, 1//1]; c = @SArray [0, 1//2];
           display(@benchmark modifying_integrator($A, $b, $c, $order))
           display(@benchmark modified_equation($A, $b, $c, $order))
       end
BenchmarkTools.Trial: 13 samples with 1 evaluation.
 Range (min … max):  405.987 ms … 419.665 ms  ┊ GC (min … max): 1.26% … 1.38%
 Time  (median):     407.149 ms               ┊ GC (median):    1.32%
 Time  (mean ± σ):   408.615 ms ±   3.711 ms  ┊ GC (mean ± σ):  1.42% ± 0.20%

  █▁▁▁▁▁▁   ▁   ▁  ▁      ▁                                   ▁  
  ███████▁▁▁█▁▁▁█▁▁█▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  406 ms           Histogram: frequency by time          420 ms <

 Memory estimate: 145.27 MiB, allocs estimate: 2973485.
BenchmarkTools.Trial: 12 samples with 1 evaluation.
 Range (min … max):  414.561 ms … 429.649 ms  ┊ GC (min … max): 1.41% … 1.56%
 Time  (median):     416.205 ms               ┊ GC (median):    1.45%
 Time  (mean ± σ):   417.728 ms ±   4.513 ms  ┊ GC (mean ± σ):  1.56% ± 0.20%

  █                                                              
  █▆▆▁▁▁▁▁▁▁▆▆▁▆▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆ ▁
  415 ms           Histogram: frequency by time          430 ms <

 Memory estimate: 145.38 MiB, allocs estimate: 2973504.
```

First commit in this PR:

```julia
julia> using RootedTrees, BSeries, BenchmarkTools, StaticArrays

julia> let order = 10
           # explicit midpoint method
           A = @SArray [0 0; 1//2 0]; b = @SArray [0, 1//1]; c = @SArray [0, 1//2];
           display(@benchmark modifying_integrator($A, $b, $c, $order))
           display(@benchmark modified_equation($A, $b, $c, $order))
       end
BenchmarkTools.Trial: 13 samples with 1 evaluation.
 Range (min … max):  390.405 ms … 407.330 ms  ┊ GC (min … max): 2.65% … 2.64%
 Time  (median):     391.723 ms               ┊ GC (median):    2.75%
 Time  (mean ± σ):   393.572 ms ±   4.565 ms  ┊ GC (mean ± σ):  2.78% ± 0.24%

  ▃  ▃█▃                                                         
  █▁▁███▁▁▁▁▁▁▁▇▁▁▇▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇ ▁
  390 ms           Histogram: frequency by time          407 ms <

 Memory estimate: 145.27 MiB, allocs estimate: 2973485.
BenchmarkTools.Trial: 13 samples with 1 evaluation.
 Range (min … max):  399.158 ms … 406.041 ms  ┊ GC (min … max): 2.61% … 2.68%
 Time  (median):     400.719 ms               ┊ GC (median):    2.63%
 Time  (mean ± σ):   401.621 ms ±   2.126 ms  ┊ GC (mean ± σ):  2.69% ± 0.21%

  ▁   ▁▁    █▁ ▁▁                █         ▁    ▁             ▁  
  █▁▁▁██▁▁▁▁██▁██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁█▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  399 ms           Histogram: frequency by time          406 ms <

 Memory estimate: 145.38 MiB, allocs estimate: 2973504.
```